tasks/mnistclass/pyproject.toml
[project]
name = "mnist-recognition"
version = "0.1.0"
description = "MNIST digit recognition with <10k parameters"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "wandb>=0.15.0",
    "hydra-core>=1.3.0",
    "ray[all,tune,wandb]>=2.5.0",
    "torchinfo>=1.8.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF"]
ignore = ["COM812", "ISC001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto" 
[tool.hatchling.build.targets.wheel] 
packages = ["."] 
